translation dashboard/requirements.txt
python-dotenv
googletrans==3.1.0a0
python-docx
pypdf
Flask